+++ outputs.tf
@@ -8,7 +8,7 @@ output "slot" {
   value       = var.instance.type == "linux" ? try(azurerm_linux_web_app_slot.linux[var.instance.name], null) : try(azurerm_windows_web_app_slot.windows[var.instance.name], null)
 }
 
-output "user_assigned_identities" {
-  description = "contains all user assigned identities configuration"
-  value       = azurerm_user_assigned_identity.identity
-}
+# output "user_assigned_identities" {
+#   description = "contains all user assigned identities configuration"
+#   value       = azurerm_user_assigned_identity.identity
+# }