+++ CHANGELOG.md
@@ -1,5 +1,22 @@
 # Changelog
 
+## [3.0.0](https://github.com/CloudNationHQ/terraform-azure-app/compare/v2.4.0...v3.0.0) (2025-04-04)
+
+
+### ⚠ BREAKING CHANGES
+
+* removed system assigned identity default
+
+### Features
+
+* add type definitions and made identity fully optional ([#31](https://github.com/CloudNationHQ/terraform-azure-app/issues/31)) ([c13fdb6](https://github.com/CloudNationHQ/terraform-azure-app/commit/c13fdb6cba5d73146bcb899d535477552349f18e))
+
+### Upgrade from v2.4.0 to v3.0.0:
+
+- Update module reference to: `version = "~> 3.0"`
+- The user assigned identity is removed from the module and it is not set to system assigned default anymore as well.
+  - For identity we created a separate module as shown in the examples.
+
 ## [2.4.0](https://github.com/CloudNationHQ/terraform-azure-app/compare/v2.3.0...v2.4.0) (2025-01-20)